app/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

google-generativeai==0.7.2

google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
